tests: support running tests on windows
