+++ panicwrap_test.go
@@ -409,22 +409,6 @@ func TestPanicWrap_panicBoundary(t *testing.T) {
 	}
 }
 
-func TestPanicWrap_monitor(t *testing.T) {
-
-	stdout := new(bytes.Buffer)
-
-	p := helperProcess("panic-monitor")
-	p.Stdout = stdout
-	//p.Stderr = new(bytes.Buffer)
-	if err := p.Run(); err == nil || err.Error() != "exit status 2" {
-		t.Fatalf("err: %s", err)
-	}
-
-	if !strings.Contains(stdout.String(), "wrapped:") {
-		t.Fatalf("didn't wrap: %#v", stdout.String())
-	}
-}
-
 func TestWrapped(t *testing.T) {
 	stdout := new(bytes.Buffer)
